Add all files
Just a side note, if you are thinking of using these for your classes, consider not doing so. The class is ever-changing and I cannot guarantee the correctness of these files and coded it with what looks like a lot of spaghetti. Also, it is not worth the Academic Dishonesty on your record, and yes, the TAs do check.